% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate-model.R
\docType{methods}
\name{agcParam}
\alias{agcParam}
\alias{agcParam,data.frame,character,character,matrix,numeric,numeric,numeric-method}
\title{Fit the covariate model for the Bayesian auto-g network computation.}
\usage{
agcParam(data, treatment, outcome, adjmat, B = 25000, R = 10,
  seed = c(1))


  \S4method{agcParam}{data.frame,character,character,matrix,numeric,numeric,numeric}(data,
  treatment, outcome, adjmat, B = 25000, R = 10, seed = c(1))
}
\arguments{
\item{data}{A data.frame containing variables to be considered
for the covariate model.}

\item{treatment}{A string specifying the column in data
that is the treatment effect in the underlying model.
Must be a binary (0/1) value.}

\item{outcome}{A string specifying the column in data
that is the outcome of interest in the underlying model.
Must be a binary (0/1) value.}

\item{adjmat}{An adjacency matrix (0/1s) specifying
the network structure. The number of rows and columns should
match the number of rows in the data object.}

\item{B}{The number of iterations for the MCMC outer loop.
Default = 25,000}

\item{R}{The number of iterations for the Gibbs inner loop.
Default = 10.}

\item{seed}{An integer vector of the values for \code{set.seed}.
By default, only c(1), which runs one chain with a seed of 1. If
the vector is of length greater than 1, then multiple chains with
both values as a seed will be run in parallel.}
}
\value{
An S3 object of type agcParamClass that contains essential
values for the covariate model.
}
\description{
\code{agcParam} takes a data frame containing a treatment,
outcome, and various covariates to be modeled via an MCMC.
The network of observations (rows in the data frame) should
be contained in the adjacency matrix (adjmat).
}
\details{
B and R specify the number of iterations to be run for the
various Bayesian sampling procedures. Defaults are conservative.

The seed variable requires a vector of integers to
run potentially multiple chains.
}
\examples{

rdsIn <- readRDS(paste0(system.file('extdata',package='autognet'),"/agc-example.rds"))
adjmat <- rdsIn[[1]]
data <- rdsIn[[2]]
treatment <- "treatment"
outcome <- "outcome"
B = 10
R = 5
seed = 1

mod <- agcParam(data, treatment, outcome, adjmat,
                         B = B, R = R, seed = c(1))

}
