% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autognet.R
\docType{package}
\name{autognet}
\alias{autognet}
\title{autognet: an R package for network data effect estimations}
\description{
The autognet R package implements a comprehensive
covariate and outcome model for the estimation of
causal direct and spill over effects from network data
}
