% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auxVarOutcomeCpp}
\alias{auxVarOutcomeCpp}
\title{Run Gibbs sampler for auxiliary outcome values using Rcpp}
\usage{
auxVarOutcomeCpp(beta, trt, cov, N, R, adjacency, start, weights)
}
\arguments{
\item{beta}{A numeric vector of parameters from outcome model}

\item{trt}{A numeric vector of the treatment values}

\item{cov}{A numeric matrix for observed covariate values (starting values for chain)}

\item{N}{An integer indicating the size of the interconnected network}

\item{R}{An integer indicating the number of iterations for the Gibbs}

\item{adjacency}{A binary matrix indicating connected units}

\item{start}{A vector of the initializing values of}

\item{weights}{A numeric vector indicating the number of neighbors for each node}
}
\value{
A numeric vector for auxiliary covariate outcomes
as an element of {0,1}
}
\description{
Given the specific inputs, determine auxiliary outcome
values using a Gibbs sampling procedure.
}
