% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{networkGibbsOutCovCpp}
\alias{networkGibbsOutCovCpp}
\title{Run Gibbs sampler for network causal effects}
\usage{
networkGibbsOutCovCpp(tau, rho, nu, ncov, R, N, rho_mat, adjacency,
  weights, cov_mat, group_lengths, group_functions)
}
\arguments{
\item{tau}{A numeric vector for the intercept terms in the covariate model}

\item{rho}{A numeric vector for the correlation terms in the covariate model}

\item{nu}{A numberic vector for the neighbor terms in the covariate model}

\item{ncov}{An integer for the number of covariates}

\item{R}{An integer indicating the number of iterations for the Gibbs}

\item{N}{An integer indicating the size of the interconnected network}

\item{rho_mat}{A numeric matrix for rho terms}

\item{adjacency}{A binary matrix indicating connected units}

\item{weights}{A numeric vector indicating the number of neighbors for each node}

\item{group_lengths}{An integer vector indicating the number of categories for each variable}

\item{group_functions}{An integer vector indicating the type of variable}

\item{cov_i}{A numeric matrix for observed covariate values (starting values for chain)}
}
\value{
A numeric matrix for auxiliary covariate values
between [0,1]
}
\description{
Given the specific inputs, determine auxiliary covariate
values using a Gibbs sampling procedure.
}
