% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{networkGibbsOuts1Cpp}
\alias{networkGibbsOuts1Cpp}
\title{IF DOCUMENT}
\usage{
networkGibbsOuts1Cpp(cov_list, beta, p, ncov, R, N, adjacency, weights,
  burnin, average)
}
\arguments{
\item{IF}{DOCUMENT}
}
\description{
Given the specific inputs, determine auxiliary outcome
values using a Gibbs sampling procedure.
}
