% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-model.R
\docType{methods}
\name{agcEffect}
\alias{agcEffect}
\alias{agcEffect,list-method}
\title{Evaluate the out model for the Bayesian auto-g network computation.}
\usage{
agcEffect(mod, burnin = 1, thin = 0.5, treatment_allocation = 0.5,
  R = 10, seed = 1)

\S4method{agcEffect}{list}(mod, burnin = 1, thin = 0.2,
  treatment_allocation = 0.5, R = 10, seed = 1)
}
\arguments{
\item{mod}{An \code{agcParamClass} object from \code{agcParam}
containing variables to be considered for the covariate model.}

\item{burnin}{The index to start evaluation as one would normally
have for a burnin for a Bayesian computation. Default = 1}

\item{thin}{The rate at which to evaluate the outcome model.
The closer to 1, the more values to compute (supplying 1 will
not thin at all).}

\item{treatment_allocation}{The proportion of the individuals
in the network to receive the treatment. Should be a number between
0 and 1. Default = 0.5.}

\item{R}{The number of iterations for the Gibbs inner loop.
Default = 10.}

\item{seed}{An integer value for \code{set.seed}.
This will be applied uniformly to all chains. Default = 1.}
}
\value{
An S3 object of type \code{agcEffectClass} that contains essential
values for the outcome model.
}
\description{
\code{agcEffect} takes a covariate model fit from \code{agcParam},
parameters for an MCMC, and the specified treatment effect and
determines the causal estimates (direct and spillover) generated
from the network structure.
}
\examples{

rdsIn <- readRDS(paste0(system.file('extdata',package='autognet'),"/agc-example.rds"))
adjmat <- rdsIn[[1]]
data <- rdsIn[[2]]
treatment <- "treatment"
outcome <- "outcome"
B = 10
R = 5
seed = 1

mod <- agcParam(data, treatment, outcome, adjmat,
                         B = B, R = R, seed = c(1))
effects <- agcEffect(mod, burnin = 1, thin = 0.5, treatment_allocation = 0.5, R = 10, seed = 1)


}
