% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal-estimates.R
\docType{methods}
\name{agcEffect}
\alias{agcEffect}
\alias{agcEffect,list-method}
\title{Evaluate the network causal effects from estimated Gibbs factors.}
\usage{
agcEffect(mod, burnin = 1, thin = 0.5, treatment_allocation = 0.5,
  subset = 0, a_fixed = c(0), dynamic_coef_vec = c(0),
  dynamic_among_treated = 0, dynamic_single_edge = 0, R = 10,
  burnin_R = 10, burnin_cov = 10, average = TRUE,
  index_override = 0, return_effects = 1, p_vec = c(0.5, 0.5))

\S4method{agcEffect}{list}(mod, burnin = 1, thin = 0.2,
  treatment_allocation = 0.5, subset = 0, a_fixed = c(0),
  dynamic_coef_vec = c(0), dynamic_among_treated = 0,
  dynamic_single_edge = 0, R = 10, burnin_R = 10, burnin_cov = 10,
  average = TRUE, index_override = 0, return_effects = 1,
  p_vec = c(0.5, 0.5))
}
\arguments{
\item{mod}{An \code{agcParamClass} object from \code{agcParam}
containing variables to be considered for the covariate model.}

\item{burnin}{The index to start evaluation as one would normally
have for a burnin for a Bayesian computation. Default = 1}

\item{thin}{The rate at which to evaluate the outcome model.
The closer to 1, the more values to compute (supplying 1 will
not thin at all).}

\item{treatment_allocation}{The proportion of the individuals
in the network to receive the treatment. Should be a number between
0 and 1. Default = 0.5.}

\item{subset}{The indices of the individuals, as they appear in the
adjacency matrix, to be included in the network causal effects estimates.
Default = 0 (include everyone).}

\item{a_fixed}{Izzie to Update
#'}

\item{dynamic_coef_vec}{Izzie to Update}

\item{dynamic_among_treated}{Izzie to Update}

\item{dynamic_single_edge}{Izzie to Update}

\item{R}{The number of iterations for the Gibbs inner loop.
Default = 10.}

\item{burnin_R}{The index to start evaluation as one would normally
have for a burnin for a Bayesian computation. Default = 10.}

\item{burnin_cov}{The index to start saving covariates for use in the
network causal effect chains. Default = 10.}

\item{average}{An indicator of whether to evaluate the causal effects as an average
of the R iterations. Default = "TRUE".}

\item{index_override}{The MCMC outer loop iteration numbers to include
in the evaluation of the causal effects. This will override the burnin and
thin parameters. Default = 0 (all iterations included).}

\item{return_effects}{An indicator of whether to return direct and spillover effects
or just the counterfactual expectations. Use the latter if you wish to construct overall
effects or different types of spillover effects. Default = 1.}

\item{p_vec}{Izzie to do}
}
\value{
An S3 object of type \code{agcEffectClass} that contains essential
values for the outcome model.
}
\description{
\code{agcEffect} takes a covariate and outcome model fit
from \code{agcParam}, parameters for an MCMC, and the
specified treatment effect and determines the causal estimates (direct and spillover)
generated from the network structure.
}
\details{
ADD DISCUSSION OF DIFFERENT TREATMENT REGIME OPTIONS by linking to URL
}
\examples{

rdsIn <- readRDS(paste0(system.file('extdata',package='autognet'),"/agc-example.rds"))
adjmat <- rdsIn[[1]]
data <- rdsIn[[2]]
treatment <- "treatment"
outcome <- "outcome"
B = 10
R = 5
seed = 1

mod <- agcParam(data, treatment, outcome, adjmat,
                         B = B, R = R, seed = c(1))
effects <- agcEffect(mod)


}
