% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-factors_DMH.R
\docType{methods}
\name{agcParam}
\alias{agcParam}
\alias{agcParam,data.frame,character,character-method}
\title{Estimate parameters from the covariate and outcome model
via Double Metropolis Hastings algorithm.}
\usage{
agcParam(data, treatment, outcome, adjmat, B = 25000, R = 10,
  seed = 1, scale.alpha = 0.005, scale.beta = 0.005,
  prior.alpha = 1, prior.beta = 1)

\S4method{agcParam}{data.frame,character,character}(data, treatment,
  outcome, adjmat, B = 25000, R = 10, seed = 1,
  scale.alpha = 0.005, scale.beta = 0.005, prior.alpha = 1,
  prior.beta = 1)
}
\arguments{
\item{data}{A data.frame containing variables to be considered
for the covariate model.}

\item{treatment}{A string specifying the column in data
that is the treatment effect in the underlying model.
Must be a binary (0/1) value.}

\item{outcome}{A string specifying the column in data
that is the outcome of interest in the underlying model.
Must be a binary (0/1) value.}

\item{adjmat}{An adjacency matrix (0/1s) specifying
the network structure. The number of rows and columns should
match the number of rows in the data object.}

\item{B}{The number of iterations for the MCMC outer loop.
Default = 25,000}

\item{R}{The number of iterations for the Gibbs inner loop.
Default = 10.}

\item{seed}{An integer value for \code{set.seed}.
By default, only \code{1}, which runs one chain with a seed of 1.}

\item{scale.alpha}{A numeric vector of values for the
parameters of the covariate model proposal distribution
variance. By default, a scalar of .005 is applied to all values. If you
wish to replace you can supply another scalar or the appropriate length
vector.}

\item{scale.beta}{A numeric vector of values for the
parameters of the outcome model proposal distribution
variance. By default, a scalar of .005 is applied to all values. If you
wish to replace you can supply another scalar or the appropriate length
vector.}

\item{prior.alpha}{A binary (0/1) indicator for the proposal distribution
for the covariate model parameters. If 1, the noninformative prior described in
the paper will be used. If 0, an improper prior (i.e. proportional to 1) will
be used. Default is 1.}

\item{prior.beta}{A binary (0/1) indicator for the proposal distribution
for the outcome model parameters. If 1, the noninformative prior described in
the paper will be used. If 0, an improper prior (i.e. proportional to 1) will
be used. Default is 1.}
}
\value{
An S3 object of type \code{agcParamClass} that contains essential
values for the covariate model.
}
\description{
\code{agcParam} takes a data frame containing a treatment,
outcome, and various covariates to be modeled via an MCMC.
The network of observations (rows in the data frame) should
be contained in the adjacency matrix (adjmat).
}
\details{
B and R specify the number of iterations to be run for the
various Bayesian sampling procedures. Defaults are conservative.

The seed variable requires a vector of integers to
run potentially multiple chains.
}
\examples{

rdsIn <- readRDS(paste0(system.file('extdata',package='autognet'),"/agc-example.rds"))
adjmat <- rdsIn[[1]]
data <- rdsIn[[2]]
treatment <- "treatment"
outcome <- "outcome"
B = 10
R = 5
seed = 1

mod <- agcParam(data, treatment, outcome, adjmat,
                         B = B, R = R, seed = c(1))

}
