% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auxVarCpp}
\alias{auxVarCpp}
\title{Calculate the Gini Index using Rcpp}
\usage{
auxVarCpp(tau, rho, nu, N, R, J, rho_mat, adjacency, cov_i, weights,
  group_lengths, group_functions)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A float of the Gini index where values are
between [0,1]
}
\description{
Given a vector of numbers, calcuate the
Gini index being sensitive to NAs
}
\examples{

x <- runif(1000)

}
